DDCap/requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.1.0
antlr4-python3-runtime==4.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
blobfile==1.2.7
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
clip @ git+https://github.com/openai/CLIP.git@40f5484c1c74edd83cb9cf687c6ab92b28d8b656
cog==0.0.3
colorama==0.4.4
cycler==0.11.0
DALL-E==0.1
Deprecated==1.2.13
docker-pycreds==0.4.0
einops==0.3.2
filelock==3.4.2
Flask==2.0.3
flatbuffers==1.12
fonttools==4.29.1
frozenlist==1.2.0
fsspec==2022.1.0
ftfy==6.0.3
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.7.0
huggingface-hub==0.4.0
idna==3.3
imageio==2.13.5
importlib-metadata==4.10.1
iniconfig==1.1.1
itsdangerous==2.1.0
Jinja2==3.0.3
joblib==1.1.0
jsonlines==3.0.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==14.0.1
lmdb==1.3.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
multidict==5.2.0
mypy==0.931
mypy-extensions==0.4.3
networkx==2.6.3
nltk==3.7
numpy==1.22.1
oauthlib==3.1.1
omegaconf==2.1.1
opencv-python-headless==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pathtools==0.1.2
Pillow==9.0.0
pluggy==1.0.0
portalocker==2.4.0
promise==2.3
protobuf==3.19.3
psutil==5.9.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocoevalcap @ git+https://github.com/jmhessel/pycocoevalcap.git@96b9f5c0e1b97642a945b3d15f5b6db5e5e138d2
cython==0.29.33
setuptools==59.6.0
pycocotools==2.0.7
pycparser==2.21
pycryptodomex==3.12.0
pyDeprecate==0.3.1
pyparsing==3.0.6
pytest==7.0.0rc1
python-dateutil==2.8.2
pytorch-lightning==1.5.8
pytz==2022.1
PyWavelets==1.2.0
PyYAML==6.0
qudida==0.0.4
redis==4.1.4
regex==2021.11.10
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
sacrebleu==2.0.0
sacremoses==0.0.47
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.5.6
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
tabulate==0.8.9
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
timm==0.3.4
tokenizers==0.11.5
tomli==2.0.0
torch==1.9.0
torchmetrics==0.6.2
torchvision==0.10.0
tqdm==4.62.3
transformers==4.16.2
typing_extensions==4.0.1
urllib3==1.26.8
wandb==0.12.10
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
xmltodict==0.12.0
yacs==0.1.8
yarl==1.7.2
yaspin==2.1.0
zipp==3.7.0
